Fix sudoku_specialized.rs not solving empty grids
If the minimum possible moves on a board is 9, ibest is never set (stays None) and the board is considered solved.

This only happens on empty grids. Since empty grids have many solutions, and none of your test cases have multiple solutions, you probably don't want to put it in anyway.